backend/requirements.txt
fastapi
uvicorn
pymupdf
langchain-core
langchain-google-genai
langchain-community
faiss-cpu
python-multipart
python-dotenv
langdetect
pydantic
pydantic-settings
google-generativeai
